Fixed ImGUI in sanbox usage, made minor fixes to the event system type macros and minor fixes to the object id generator

+++ WCCEngine/src/Engine/Core/MouseCodes.h
@@ -3,7 +3,7 @@
 
 namespace WCCEngine
 {
-	enum MouseCodes : short
+	enum MouseCodes : int
 	{
 		Button0 = 0,
 		Button1 = 1,
